+++ pom.xml
@@ -43,6 +43,11 @@
             <artifactId>mssql-jdbc</artifactId>
             <version>12.8.1.jre8</version>
         </dependency>
+        <dependency>
+            <groupId>org.postgresql</groupId>
+            <artifactId>postgresql</artifactId>
+            <version>42.7.7</version>
+        </dependency>
     </dependencies>
 
     <build>
