+++ src/main/java/com/cinemax/empleados/Main.java
@@ -1,11 +1,14 @@
 package com.cinemax.empleados;
 
+import com.cinemax.empleados.comun.ConexionBaseSingleton;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 
 import java.io.IOException;
+import java.sql.Connection;
+import java.sql.SQLException;
 
 
 public class Main extends Application {
@@ -13,7 +16,7 @@ public class Main extends Application {
 
     @Override
     public void start(Stage stage) throws IOException {
-                
+
         FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/Vista/empleados/PantallaLogin.fxml"));
         Scene scene = new Scene(fxmlLoader.load(), 800, 500);
 
@@ -22,10 +25,29 @@ public class Main extends Application {
         stage.setScene(scene);
         stage.show();
     }
-
+    @Override
+    public void stop() {
+        // Aquí cierras la conexión
+        System.out.println("Cerrando conexion");
+        ConexionBaseSingleton.getInstancia().cerrar();
+    }
 
 
     public static void main(String[] args) {
+        try {
+            /* Carga explícita del driver JDBC */
+            Class.forName("org.postgresql.Driver");
+
+            /* Ahora sí pedimos la conexión al Singleton */
+            try {
+                ConexionBaseSingleton.getInstancia().getConexion();
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+            System.out.println("¡Conexión exitosa!");
+        } catch (Exception e) {
+            System.err.println("No se encontró el driver PostgreSQL: " + e.getMessage());
+        }
         // You can add any pre-launch logic here if needed
         launch();
     }
