+++ src/main/java/com/cinemax/empleados/modelo/Entidades/Rol.java
@@ -1,4 +1,4 @@
-package com.cinemax.empleados.Modelo.Entidades;
+package com.cinemax.empleados.modelo.Entidades;
 
 import java.util.Set;
 import java.util.HashSet;
