+++ src/main/java/com/cinemax/empleados/modelo/Entidades/Usuario.java
@@ -1,4 +1,4 @@
-package com.cinemax.empleados.Modelo.Entidades;
+package com.cinemax.empleados.modelo.Entidades;
 
 import java.time.LocalDateTime;
