+++ src/main/java/com/cinemax/empleados/servicios/ServicioRoles.java
@@ -1,8 +1,8 @@
-package com.cinemax.empleados.Servicios;
+package com.cinemax.empleados.servicios;
 
-import com.cinemax.empleados.Modelo.Entidades.Rol;
-import com.cinemax.empleados.Modelo.Entidades.Permiso;
-import com.cinemax.empleados.Modelo.Persistencia.RolDAO;
+import com.cinemax.empleados.modelo.Entidades.Rol;
+import com.cinemax.empleados.modelo.Entidades.Permiso;
+import com.cinemax.empleados.modelo.Persistencia.RolDAO;
 import java.util.List;
 import java.util.Set;
