+++ src/main/java/module-info.java
@@ -5,8 +5,8 @@ module com.cinemax.empleados {
 
 
 opens com.cinemax.empleados to javafx.fxml;
-opens com.cinemax.empleados.Modelo.Entidades to javafx.base;
+opens com.cinemax.empleados.modelo.Entidades to javafx.base;
 exports com.cinemax.empleados;
-opens com.cinemax.empleados.Controlador to javafx.fxml;
-exports com.cinemax.empleados.Controlador;
+opens com.cinemax.empleados.controlador to javafx.fxml;
+exports com.cinemax.empleados.controlador;
 }
