+++ src/main/resources/Vista/empleados/PantallaPortalPrincipal.fxml
@@ -7,7 +7,7 @@
 
 <!-- prefHeight="40" -->
 
-<BorderPane prefHeight="550.0" prefWidth="800.0" stylesheets="@ayu-theme.css" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.cinemax.empleados.Controlador.ControladorPortalPrincipal">
+<BorderPane prefHeight="550.0" prefWidth="800.0" stylesheets="@ayu-theme.css" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.cinemax.empleados.controlador.ControladorPortalPrincipal">
     <top>
         <HBox fx:id="headerBar" alignment="CENTER_LEFT" spacing="15.0" styleClass="header-bar">
             <children>
